added animated emoji support

+++ bot.py
@@ -21,7 +21,7 @@ help_message = [
     ['help', 'Display this glorious help message.']
 ]
 status = {"maintain_emoji_state": "idle"}
-emoji_id_dict = {}
+emoji_dict = {}
 
 class Paginator:
     def __init__(self, items, per_page=10):
@@ -57,36 +57,83 @@ class MyClient(discord.Client):
         for i in duplicates:
             emoji = discord.utils.get(guild_emojis, name=i)
             await emoji.delete()
-        return set(guild_emoji_names)
+
+        result = {'regular': set(), 'animated': set()}
+        guild = client.get_guild(guild_id)
+        guild_emojis = guild.emojis
+        for i in guild_emojis:
+            if i.animated:
+                result['animated'].add(i.name)
+            else:
+                result['regular'].add(i.name)
+        return result
 
     async def set_guild_emoji_state(self, guild_id, bucket_name, new_state):
         guild = client.get_guild(guild_id)
         guild_state = await self.get_guild_emoji_state(guild_id)
-        to_delete = guild_state - new_state
-        to_add = new_state - guild_state
-        for i in to_delete:
-            await discord.utils.get(guild.emojis, name=i).delete()
-        for i in to_add:
-            emoji_data = minioClient.get_object(bucket_name, i).read()
-            emoji = await guild.create_custom_emoji(name=i, image=emoji_data)
-            emoji_id_dict[emoji.name] = emoji.id
+
+        if type(new_state) == type(set()):
+            bucket_state = await self.get_bucket_emoji_state(bucket_name)
+            new_state_dict = {'animated': set(), 'regular': set()}
+            for i in new_state:
+                if i in bucket_state['regular']:
+                    new_state_dict['regular'].add(i)
+                elif i in new_state_dict['animated']:
+                    new_state_dict['animated'].add(i)
+            new_state = new_state_dict
+
+        if "regular" in new_state:
+            to_delete_regular = (guild_state['regular'] - new_state['regular'])
+            to_add_regular = (new_state['regular'] - guild_state['regular'])
+            for i in to_delete_regular:
+                await discord.utils.get(guild.emojis, name=i).delete()
+            for i in to_add_regular:
+                emoji_data = minioClient.get_object(bucket_name, "regular/" + i).read()
+                emoji = await guild.create_custom_emoji(name=i, image=emoji_data)
+                emoji_dict[emoji.name] = [emoji.id, emoji.animated]
+
+        if "animated" in new_state:
+            to_delete_animated = (guild_state['animated'] - new_state['animated'])
+            to_add_animated = (new_state['animated'] - guild_state['animated'])
+            for i in to_delete_animated:
+                await discord.utils.get(guild.emojis, name=i).delete()
+            for i in to_add_animated:
+                emoji_data = minioClient.get_object(bucket_name, "animated/" + i).read()
+                emoji = await guild.create_custom_emoji(name=i, image=emoji_data)
+                emoji_dict[emoji.name] = [emoji.id, emoji.animated]
 
     async def get_bucket_emoji_state(self, bucket_name):
-        return set([i.object_name for i in minioClient.list_objects_v2(bucket_name)])
+        result = {}
+        result['regular'] = set([i.object_name.replace("regular/", "") for i in minioClient.list_objects_v2(bucket_name, prefix="regular/") if not i.object_name == 'regular/'])
+        result['animated'] = set([i.object_name.replace("animated/", "") for i in minioClient.list_objects_v2(bucket_name, prefix="animated/") if not i.object_name == 'animated/'])
+        return result
 
     async def set_bucket_emoji_state(self, guild_id, bucket_name, new_state):
         guild = client.get_guild(guild_id)
         bucket_state = await self.get_bucket_emoji_state(bucket_name)
-        to_delete = bucket_state - new_state
-        to_add = new_state - bucket_state
-        minioClient.remove_objects(bucket_name, to_delete)
-        to_add_emoji_objects = [i for i in guild.emojis if i.name in to_add]
+        to_add_emoji_objects = []
+
+        if "regular" in new_state:
+            to_delete_regular = bucket_state['regular'] - new_state['regular']
+            to_add_regular = new_state['regular'] - bucket_state['regular']
+            minioClient.remove_objects(bucket_name, ["regular/" + i for i in to_delete_regular])
+
+            to_add_emoji_objects.extend([i for i in guild.emojis if i.name in to_add_regular])
+        if "animated" in new_state:
+            to_delete_animated = bucket_state['animated'] - new_state['animated']
+            to_add_animated = new_state['animated'] - bucket_state['animated']
+            minioClient.remove_objects(bucket_name, ['animated/' + i for i in to_delete_animated])
+            to_add_emoji_objects.extend([i for i in guild.emojis if i.name in to_add_animated])
+
         for i in to_add_emoji_objects:
             emoji_data = io.BytesIO(await i.url.read())
             emoji_data.seek(0, 2)
             emoji_data_len = emoji_data.tell()
             emoji_data.seek(0, 0)
-            minioClient.put_object(bucket_name, i.name, emoji_data, length=emoji_data_len, content_type="image/png")
+            if i.animated:
+                minioClient.put_object(bucket_name, "animated/" + i.name, emoji_data, length=emoji_data_len, content_type="image/gif")
+            else:
+                minioClient.put_object(bucket_name, "regular/" + i.name, emoji_data, length=emoji_data_len, content_type="image/png")
 
 
     async def maintain_emoji_state(self, guild_id, bucket_name):
@@ -96,44 +143,63 @@ class MyClient(discord.Client):
         status['maintain_emoji_state'] = time.time()
 
         guild = client.get_guild(guild_id)
-        guild_state = await self.get_guild_emoji_state(guild_id)
-        bucket_state = await self.get_bucket_emoji_state(bucket_name)
+        total_guild_state = await self.get_guild_emoji_state(guild_id)
+        total_bucket_state = await self.get_bucket_emoji_state(bucket_name)
 
-        if guild_state == bucket_state:
-            return
-        if not guild_state.issubset(bucket_state):
-            await self.set_bucket_emoji_state(guild_id, bucket_name, bucket_state.union(guild_state))
-        if len(guild_state) < min(len(bucket_state), guild.emoji_limit - 1):
-            num_needed_guild_emojis = min(len(bucket_state), guild.emoji_limit - 1) - len(guild_state)
-            list_bucket_unique_state = list(bucket_state - guild_state)
-            random.shuffle(list_bucket_unique_state)
-            await self.set_guild_emoji_state(guild_id, bucket_name, guild_state.union(set(list_bucket_unique_state[:num_needed_guild_emojis])))
-        if len(guild_state) == guild.emoji_limit:
-            guild_emojis = list(guild.emojis)
-            guild_emojis.sort(key=lambda x: x.created_at)
-            await self.set_guild_emoji_state(guild_id, bucket_name, guild_state - {guild_emojis[0].name})
+        for emoji_type in ['regular', 'animated']:
+            guild_state = total_guild_state[emoji_type]
+            bucket_state = total_bucket_state[emoji_type]
+
+            if guild_state == bucket_state:
+                continue
+            if not guild_state.issubset(bucket_state):
+                await self.set_bucket_emoji_state(guild_id, bucket_name, {emoji_type: bucket_state.union(guild_state)})
+            if len(guild_state) < min(len(bucket_state), guild.emoji_limit - 1):
+                num_needed_guild_emojis = min(len(bucket_state), guild.emoji_limit - 1) - len(guild_state)
+                list_bucket_unique_state = list(bucket_state - guild_state)
+                random.shuffle(list_bucket_unique_state)
+                await self.set_guild_emoji_state(guild_id, bucket_name, {emoji_type: guild_state.union(set(list_bucket_unique_state[:num_needed_guild_emojis]))})
+            if len(guild_state) == guild.emoji_limit:
+                guild_emojis = list(guild.emojis)
+                guild_emojis.sort(key=lambda x: x.created_at)
+                await self.set_guild_emoji_state(guild_id, bucket_name, {emoji_type: guild_state - {guild_emojis[0].name}})
 
         status['maintain_emoji_state'] = "idle"
 
     def sub_emoji(self, matchObj):
-        emoji_id = emoji_id_dict[matchObj.group(1)]
-        return "<:{0}:{1}>".format(matchObj.group(1), emoji_id)
+        emoji = emoji_dict[matchObj.group(1)]
+        emoji_id = emoji[0]
+        emoji_animated = emoji[1]
+        if emoji_animated:
+            return "<a:{0}:{1}>".format(matchObj.group(1), emoji_id)
+        else:
+            return "<:{0}:{1}>".format(matchObj.group(1), emoji_id)
 
     async def delete_emoji(self, guild_id, bucket_name, emoji_name, edit_guild=True):
+        guild = client.get_guild(guild_id)
         if edit_guild:
-            guild = client.get_guild(guild_id)
-            emoji = await discord.utils.get(guild.emojis, name=curName)
+            emoji = discord.utils.get(guild.emojis, name=curName)
             await emoji.delete()
-        minioClient.remove_object(bucket_name, emoji_name)
+        if emoji_dict[emoji_name][1]:
+            obj_prefix = "animated/"
+        else:
+            obj_prefix = "regular/"
+        minioClient.remove_object(bucket_name, obj_prefix + emoji_name)
 
     async def rename_emoji(self, guild_id, bucket_name, curName, newName, edit_guild=True):
+        guild = client.get_guild(guild_id)
         if edit_guild:
-            guild = client.get_guild(guild_id)
-            emoji = await discord.utils.get(guild.emojis, name=curName)
+            emoji = discord.utils.get(guild.emojis, name=curName)
             await emoji.edit(name="newName")
-        curObj = minioClient.stat_object(bucket_name, curName)
+        else:
+            emoji = discord.utils.get(guild.emojis, name=newName)
+        if emoji.animated:
+            obj_prefix = "animated/"
+        else:
+            obj_prefix = "regular/"
+        curObj = minioClient.stat_object(bucket_name, obj_prefix + curName)
         metadata = curObj.metadata
-        minioClient.copy_object(bucket_name, newName, '{0}/{1}'.format(bucket_name, curName), metadata=metadata)
+        minioClient.copy_object(bucket_name, obj_prefix + newName, '{0}/{1}'.format(bucket_name, obj_prefix + curName), metadata=metadata)
         await self.delete_emoji(guild_id, bucket_name, curName, edit_guild=False)
 
     async def on_ready(self):
@@ -143,7 +209,7 @@ class MyClient(discord.Client):
 
         self.guild = await client.fetch_guild(config_dict['guild_id'])
         for i in self.guild.emojis:
-            emoji_id_dict[i.name] = i.id
+            emoji_dict[i.name] = [i.id, i.animated]
         await self.maintain_emoji_state(config_dict['guild_id'], config_dict['bucket_name'])
 
     async def on_guild_emojis_update(self, guild, before, after):
@@ -179,27 +245,30 @@ class MyClient(discord.Client):
                         page = 0
                     else:
                         page = int(command[1]) - 1
-                    objects = minioClient.list_objects_v2(config_dict['bucket_name'])
-                    paginator = Paginator(list(objects), 50)
+                    bucket_state = await self.get_bucket_emoji_state(config_dict['bucket_name'])
+                    all_emojis = bucket_state['regular'].union(bucket_state['animated'])
+                    paginator = Paginator(list(all_emojis), 50)
                     num_pages = paginator.num_pages()
-                    objects = paginator.get_page(page)
+                    emojis = paginator.get_page(page)
                     next_page = (page+2)%num_pages
                     if next_page == 0:
                         next_page = num_pages
                     guild_state = await self.get_guild_emoji_state(config_dict['guild_id'])
+                    all_guild_emojis = guild_state['regular'].union(guild_state['animated'])
                     embed = discord.Embed(title="Emojis of {0}".format(message.guild.name))
                     emoji_textlist = ['**__Page {0}/{1}__**\n'.format(page+1, num_pages)]
-                    for i in objects:
+                    for i in emojis:
                         # if (not None == i.metadata) and 'x-amz-meta-info' in i.metadata:
                         #     info = i.metadata['x-amz-meta-info']
                         # else:
                         #     info = '*Last Modified at {0}*'.format(i.last_modified)
                         #     info = "*No info provided*"
                         # emoji_textlist.append("{0} - {1}".format(i.object_name, info))
-                        if i.object_name in guild_state:
-                            obj_name = '**{0}**'.format(i.object_name)
-                        else:
-                            obj_name = '{0}'.format(i.object_name)
+                        obj_name = '{0}'.format(i)
+                        if i in all_guild_emojis:
+                            obj_name = '**{0}**'.format(obj_name)
+                        if i in bucket_state['animated']:
+                            obj_name = '*{0}*'.format(obj_name)
                         emoji_textlist.append(obj_name)
                     emoji_textlist.append('\n**__Page {0}/{1}__**'.format(page+1, num_pages))
                     emoji_textlist.append('*Type `{0}l {1}` to view the next page*'.format(config_dict['prefix'], next_page))
@@ -234,23 +303,38 @@ class MyClient(discord.Client):
                     await self.maintain_emoji_state(config_dict['guild_id'], config_dict['bucket_name'])
             else:
                 emoji_matches = set([i for i in emoji_regex.findall(message.content)])
-                guild_emojis = set([i.name for i in await message.guild.fetch_emojis()])
-                absent_emojis = emoji_matches - guild_emojis
-                if len(absent_emojis) > 0:
-                    present_emojis = emoji_matches.union(guild_emojis)
-                    other_emojis = guild_emojis - present_emojis
-                    other_emojis_list = list(other_emojis)
-                    random.shuffle(other_emojis_list)
-                    set_emoji_coro = asyncio.create_task(self.set_guild_emoji_state(config_dict['guild_id'], config_dict['bucket_name'], absent_emojis.union(present_emojis).union(set(other_emojis_list[:len(other_emojis_list) - len(absent_emojis)]))))
-                    maintain_emoji_coro = asyncio.create_task(self.maintain_emoji_state(config_dict['guild_id'], config_dict['bucket_name']))
-                    done, pending = await asyncio.wait([set_emoji_coro, maintain_emoji_coro], return_when=asyncio.ALL_COMPLETED)
-                    if not set_emoji_coro in done:
-                        return
-                    new_message = re.sub(emoji_regex, self.sub_emoji, message.content)
-                    webhook = await message.channel.create_webhook(name=message.author.display_name, avatar=await message.author.avatar_url.read())
-                    await webhook.send(new_message)
-                    await webhook.delete()
-                    await message.delete()
+                if len(emoji_matches) > 0:
+                    guild_state = await self.get_guild_emoji_state(config_dict['guild_id'])
+                    bucket_state = await self.get_bucket_emoji_state(config_dict['bucket_name'])
+
+                    used_regular_emojis = set([i for i in emoji_matches if i in bucket_state['regular']])
+                    absent_regular_emojis = used_regular_emojis - guild_state['regular']
+                    present_regular_emojis = used_regular_emojis.union(guild_state['regular'])
+                    other_regular_emojis = guild_state['regular'] - present_regular_emojis
+                    other_regular_emojis_list = list(other_regular_emojis)
+                    random.shuffle(other_regular_emojis_list)
+                    desired_other_regular_emojis = set(other_regular_emojis_list[:len(other_regular_emojis) - len(absent_regular_emojis)])
+                    desired_regular_emoji_state = absent_regular_emojis.union(present_regular_emojis).union(desired_other_regular_emojis)
+                    
+                    used_animated_emojis = set([i for i in emoji_matches if i in bucket_state['animated']])
+                    absent_animated_emojis = used_animated_emojis - guild_state['animated']
+                    present_animated_emojis = used_animated_emojis.union(guild_state['animated'])
+                    other_animated_emojis = guild_state['animated'] - present_animated_emojis
+                    other_animated_emojis_list = list(other_animated_emojis)
+                    random.shuffle(other_animated_emojis_list)
+                    desired_other_animated_emojis = set(other_animated_emojis_list[:len(other_animated_emojis) - len(absent_regular_emojis)])
+                    desired_animated_emoji_state = absent_animated_emojis.union(present_animated_emojis).union(desired_other_animated_emojis)
+
+                    total_desired_state = {'regular': desired_regular_emoji_state, 'animated': desired_animated_emoji_state}
+                    await self.set_guild_emoji_state(config_dict['guild_id'], config_dict['bucket_name'], total_desired_state)
+
+                    if len(absent_regular_emojis) > 0 or len(used_animated_emojis) > 0:
+                        await self.maintain_emoji_state(config_dict['guild_id'], config_dict['bucket_name'])
+                        new_message = re.sub(emoji_regex, self.sub_emoji, message.content)
+                        webhook = await message.channel.create_webhook(name=message.author.display_name, avatar=await message.author.avatar_url.read())
+                        await webhook.send(new_message)
+                        await webhook.delete()
+                        await message.delete()
 
 
 client = MyClient()
